+++ zk-keygen/src/main.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use {
     bip39::{Mnemonic, MnemonicType, Seed},
     clap::{crate_description, crate_name, Arg, ArgMatches, Command},
